move static-specific write fns into build mode
